Adds new backend atomics
This patch ports a patchset by @SakalisC to ArgoDSM.

The new backend atomics differ from the previous one in two aspects:

1) The templated code has been moved to the common header file, and each
backend implements a non-templated, type-erased version of the
functions. This means that we no longer need to explicitly instantiate
every single atomic function for every single type we use it for. The
type erasure is not a very big disadvantage because the MPI backend
(which is the main backend) requires it anyway.

2) There are now more functions. Specifically, we now have atomic exchange,
store, load, CAS, and fetch&add functions.

In addition to the changes mentioned above, some tests have been added,
as well as a copy constructor to the global_ptr class.