Switch from pandoc to Jekyll
Pandoc is not easy to use because the constant changes in its output
style force us to use a very specific pandoc version. Jekyll on the
other hand not only allows us to ditch the html files and use just
markdown, but it is also the recommended way for publishing markdown
sites on Github.