Run tests for the MPI backend through mpirun (#38)
* Add option to run make test on multiple nodes

This commit ensures that make test now executes each MPI test
through mpirun. The number of software nodes to execute with make
test is set through the CMake option ARGO_TESTS_NPROCS (1-8),
with the default value being 2.

* Update documentation with make test changes

The documentation now mentions that make test executes all MPI
tests on two ArgoDSM software nodes, and that this number can be
changed through setting a CMake option.

* Ensure that tests work with multiple nodes

This commit ensures that ArgoDSM is initialized using a pre-defined
cache size in addition to memory size. In tests using a larger memory
size, ArgoDSM is initialized with a significantly smaller cache size
to ensure that the test does not use too much local memory.