Improved ArgoDSM write buffer
This commit adds a dedicated ArgoDSM write buffer that now acts
as the base for deciding whether or not to write back pages
instead of iterating over the entire cache checking for dirty
pages. It also adds a foundation for selective coherence correctly
removing downgraded pages from the write buffer.

The write buffer temporarily provides functionality to write back
all, or a number of, the ArgoDSM pages corresponding to the
indices in the write buffer. This should eventually be replaced by
a dedicated cache module with an appropriate API.

The environment variables ARGO_WRITE_BUFFER_SIZE and
ARGO_WRITE_BUFFER_WRITE_BACK_SIZE can be used to set the
maximum number of ArgoDSM pages contained in the write buffer and
the number of pages to write back once the write buffer is full
respectively. The maximum size defaults to 512 pages, and the
write back size defaults to 32 pages.