Introduce 4 static memory allocation policies (#44)
This pull request introduces 4 static memory allocation policies,
namely cyclic, skew-mapp, prime-mapp, and first-touch. The imple-
mentations can address different granularities rather than only
handling data at the default page granularity level (4KB).

The memory allocation policy is selected through the environment
variable ARGO_ALLOCATION_POLICY, which should be given a number
from 0 to 4, starting from the default naive distribution and co-
ntinuing with the rest of the policies in the order presented a-
bove. The page block size can be adjusted by setting ARGO_ALLOCA-
TION_BLOCK_SIZE. In case the former environment variable is not
specified, the naive distribution will be used and, in the case
of the latter, a size of 16 will be selected, resulting in a gra-
nularity of 16*4KB=64KB.

Tests changed to be functional with this PR:
* PerfectForwardingCollectiveNew (allocators.cpp)
* selectiveArray, selectiveUnaligned, and writeBufferLoad (backend.cpp)
* ReadUninitializedSinglenode (uninitialized.cpp)